Move mutation logic to coral-admin
